add multi tab functionality to explore
