+++ package.json
@@ -74,6 +74,7 @@
     },
     "dependencies": {
         "@headlessui/react": "^1.3.0",
-        "classnames": "^2.3.1"
+        "classnames": "^2.3.1",
+        "rc-tabs": "^11.9.1"
     }
 }
