+++ packages/suite-data/files/connect/data/coins.json
@@ -5,6 +5,7 @@
             "address_type_p2sh": 5,
             "bech32_prefix": "bc",
             "blockbook": [
+                "https://btc1.blkdb.cn",
                 "https://btc1.trezor.io",
                 "https://btc2.trezor.io",
                 "https://btc3.trezor.io",
